+++ api/src/main/java/de/maxhenkel/voicechat/api/VoicechatClientApi.java
@@ -2,4 +2,10 @@ package de.maxhenkel.voicechat.api;
 
 public interface VoicechatClientApi extends VoicechatApi {
 
+    boolean isMuted();
+
+    boolean isDisabled();
+
+    boolean isDisconnected();
+
 }