+++ common/src/main/java/de/maxhenkel/voicechat/voice/client/ClientPlayerStateManager.java
@@ -2,6 +2,9 @@ package de.maxhenkel.voicechat.voice.client;
 
 import de.maxhenkel.voicechat.Voicechat;
 import de.maxhenkel.voicechat.VoicechatClient;
+import de.maxhenkel.voicechat.api.events.ClientVoicechatConnectionEvent;
+import de.maxhenkel.voicechat.api.events.MicrophoneMuteEvent;
+import de.maxhenkel.voicechat.api.events.VoicechatDisableEvent;
 import de.maxhenkel.voicechat.gui.CreateGroupScreen;
 import de.maxhenkel.voicechat.gui.EnterPasswordScreen;
 import de.maxhenkel.voicechat.gui.GroupScreen;
@@ -10,6 +13,11 @@ import de.maxhenkel.voicechat.intercompatibility.ClientCompatibilityManager;
 import de.maxhenkel.voicechat.intercompatibility.CommonCompatibilityManager;
 import de.maxhenkel.voicechat.net.NetManager;
 import de.maxhenkel.voicechat.net.UpdateStatePacket;
+import de.maxhenkel.voicechat.plugins.PluginManager;
+import de.maxhenkel.voicechat.plugins.impl.VoicechatClientApiImpl;
+import de.maxhenkel.voicechat.plugins.impl.events.ClientVoicechatConnectionEventImpl;
+import de.maxhenkel.voicechat.plugins.impl.events.MicrophoneMuteEventImpl;
+import de.maxhenkel.voicechat.plugins.impl.events.VoicechatDisableEventImpl;
 import de.maxhenkel.voicechat.voice.common.ClientGroup;
 import de.maxhenkel.voicechat.voice.common.PlayerState;
 import net.minecraft.ChatFormatting;
@@ -73,6 +81,7 @@ public class ClientPlayerStateManager {
     public void onVoiceChatDisconnected() {
         disconnected = true;
         syncOwnState();
+        PluginManager.instance().dispatchEvent(ClientVoicechatConnectionEvent.class, new ClientVoicechatConnectionEventImpl(new VoicechatClientApiImpl(), false));
     }
 
     /**
@@ -81,6 +90,7 @@ public class ClientPlayerStateManager {
     public void onVoiceChatConnected(ClientVoicechatConnection client) {
         disconnected = false;
         syncOwnState();
+        PluginManager.instance().dispatchEvent(ClientVoicechatConnectionEvent.class, new ClientVoicechatConnectionEventImpl(new VoicechatClientApiImpl(), true));
     }
 
     private void onDisconnect() {
@@ -129,6 +139,7 @@ public class ClientPlayerStateManager {
     public void setDisabled(boolean disabled) {
         VoicechatClient.CLIENT_CONFIG.disabled.set(disabled).save();
         syncOwnState();
+        PluginManager.instance().dispatchEvent(VoicechatDisableEvent.class, new VoicechatDisableEventImpl(new VoicechatClientApiImpl(), disabled));
     }
 
     public boolean isDisconnected() {
@@ -141,6 +152,7 @@ public class ClientPlayerStateManager {
 
     public void setMuted(boolean muted) {
         VoicechatClient.CLIENT_CONFIG.muted.set(muted).save();
+        PluginManager.instance().dispatchEvent(MicrophoneMuteEvent.class, new MicrophoneMuteEventImpl(new VoicechatClientApiImpl(), muted));
     }
 
     public boolean isInGroup() {