+++ fabric/changelog.md
@@ -1,2 +1,1 @@
-- Added client audio data override to API
-- Updated to 22w05a
+- Added client state events to API